musiccreation/backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Bcrypt==1.0.1
Pillow==11.2.1
python-dotenv==1.1.1
Werkzeug==2.3.7
mysql-connector-python==9.4.0 
